% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accidents.R
\docType{data}
\name{accidents}
\alias{accidents}
\title{Traffic accident rate.}
\format{A data frame 16 observations and four variables:
\describe{
   \item{accidents:}{number of accidents}
   \item{years:}{number of years of observation}
   \item{location:}{monitored location (one of eight)}
   \item{treatment:}{before or after introduction of traffic control measures}
}}
\source{
Pawitan, Y. (2013). \emph{In all likelihood: Statistical modeling and inference using likelihood}. Oxford University Press.
}
\usage{
accidents
}
\description{
Data from an observational study that recorded the number of accidents at eight locations before and after installation of traffic control measures.
}
\note{
The data are from an example from Pawitan (2013). No other source is given so the data are assumed to be artificial.
}
\keyword{datasets}
