% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acetazolamide.R
\docType{data}
\name{acetazolamide}
\alias{acetazolamide}
\title{Effect of acetazolamide on forelimb malformation.}
\format{A data frame with 588 observations and two variables:
\describe{
   \item{dose:}{dose (mg/kg)}
   \item{limbs:}{number of malformed limbs}
}}
\source{
L. B. Holmes, Kawanishi, H., & Munz, A. (1988). Acetazolamide: Maternal toxicity, pattern of malformations, and litter effect. \emph{Taratology}, \emph{37}, 335-342.

Ramsey, F. L. & Schafer, D. W. (1997). \emph{The statistical sleuth: A course in methods of data analysis}. Belmont, CA: Duxbury.
}
\usage{
acetazolamide
}
\description{
Data from an experiment investigating the effect of acetazolamide on forelimb malformation in mice. Mouse fetuses were injected with acetazolamide on the 9th day of gestation. The number of malformed forelimbs for each fetus was observed on the 18th day of gestation.
}
\note{
The fetuses may have been nested within litters. Litter membership of each fetus, however, is unknown. The data are from Holmes, Kawanishi, and Munz (1988) and are featured in Ramsey and Schafer (1997).
}
\keyword{datasets}
