% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfail.R
\docType{data}
\name{acfail}
\alias{acfail}
\title{Failures of air conditioning equipment.}
\format{
A data frame of 213 observations on two variables.
\describe{
   \item{aircraft:}{identifier of the aircraft}
   \item{time:}{time until failure (hours)}
}
}
\source{
Gaver, D. P. & O'Muircheartaigh, I. G. (1987). Robust empirical Bayes analysis of event rates. \emph{Technometrics}, \emph{29}, 1--15.

Proschan, F. (1963). Theoretical explanation of observed decreasing failure rate. \emph{Technometrics}, \emph{5}, 375--383.
}
\usage{
acfail
}
\description{
Successive times until failure of air conditioning equipment on Boeing 720 aircraft.
}
\note{
The data are from Proschan (1963) who reported the times between successive failures on each plane. Gaver and O'Muircheartaigh (1987) featured the same data but as the number of failures for each aircraft.
}
\keyword{datasets}
