% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acidrain.R
\docType{data}
\name{acidrain}
\alias{acidrain}
\title{Acid rain and tree roots.}
\format{A data frame 168 observations and 15 variables:
\describe{
   \item{lysimeter:}{number identifier of lysimeter used (5, 6, 7, or 8)}
   \item{condition:}{treatment condition (control or simulated acid rain)}
   \item{type:}{type of mycorrhiza}
   \item{sector:}{sector of the lysimeter}
   \item{0-3:}{count at depth of 0-3 cm}
   \item{3-6:}{count at depth of 3-6 cm}
   \item{6-9:}{count at depth of 6-9 cm}
   \item{9-12:}{count at depth of 9-12 cm}
   \item{12-15:}{count at depth of 12-15 cm}
   \item{15-18:}{count at depth of 15-18 cm}
   \item{18-21:}{count at depth of 18-21 cm}
   \item{21-24:}{count at depth of 21-24 cm}
   \item{24-27:}{count at depth of 24-27 cm}
   \item{27-30:}{count at depth of 27-30 cm}
   \item{30-33:}{count at depth of 30-33 cm}
}}
\source{
Jeffers, J. N. R. (1987). Acid rain and tree roots: An analysis of an experiment. In D. J. Hand & B. S. Everitt (Eds.), \emph{The statistical consultant in action} (80-107). Cambridge: Cambridge University Press.
}
\usage{
acidrain
}
\description{
Data from an experiment investigating the effect of acid rain on tree roots. Four lysimeters were assigned to control and acid treatment conditions. Fine tree roots were counted at eleven depths from cores sampled within six sectors in each lysimeter. Counts were made separately for each of seven types of mycorrhizas.
}
\note{
See Jeffers (1987) for details about the experimental design and motivation for the study.
}
\keyword{datasets}
