% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algalinhibit.R
\docType{data}
\name{algalinhibit}
\alias{algalinhibit}
\title{Effect of cadmium on algal biomass/growth.}
\format{A data frame with 68 observations and two variables:
\describe{
   \item{concentration}{concentration of cadmium in micrograms per liter}
   \item{biomass}{algal biomass (cells per mm)}
}}
\source{
Bailer, A. J. & Oris, J. T. (1997). Estimating inhibition concentrations for differing response scales using generalized linear models. \emph{Environmental Toxicology and Chemistry}, \emph{16}, 1554-1559.
}
\usage{
algalinhibit
}
\description{
Data from an experiment investigating the effect of cadmium exposure of 96 hours at three concentrations on the inhibition of algal (Selenastrum capricornutum) boimass/growth.
}
\keyword{datasets}
