% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Optimum allocation for stratified random sampling.}
\source{
Cochran, W. G. (1977). \emph{Sampling techniques} (3rd Edition). New York: Wiley.
}
\usage{
allocate(Ni, si, ci = rep(1, length(Ni)), c0 = 0, ct = NA, ev = NA)
}
\arguments{
\item{Ni}{Vector of total number of sampling units in each stratum.}

\item{si}{Vector of the standard deviation for each statum}

\item{ci}{Vector of the cost per unit for each stratum.}

\item{c0}{Overhead survey cost.}

\item{ct}{Total cost for the survey.}

\item{ev}{Variance for the estimator of the population mean.}
}
\value{
A list of a vector of the proportion of units to sample from each stratum, a vector of sample sizes for each stratum, total sample size, estimator variance, and total survey cost. If neither the estimator variance (\code{ev}) or the total cost (\code{ct}) are specified then only the proportions can be computed.
}
\description{
This is a function to compute the optimum allocation for a stratified random sampling design.
}
\details{
The solution is based choosing the overall sample size and the sample sizes for each stratum to minimize the estimator variance for a fixed cost, or to minimize the cost for a fixed estimator variance. The solution assumes the usual estimator for the for the population mean and a variance based on a finite sample design-based approach. The total cost of the survey is assumed to be the overhead cost plus a per-unit cost for each sampled unit. See Cochran (1977) for details.
}
\examples{
# sampling fractions only
allocate(Ni = c(155,62,93), si = c(5,15,10), ci = c(9,9,16))

# allocation for estimator variance fixed at 1 
allocate(Ni = c(155,62,93), si = c(5,15,10), ci = c(9,9,16), ev = 1)

# allocation for total survey cost fixed at 500
allocate(Ni = c(155,62,93), si = c(5,15,10), ci = c(9,9,16), ct = 500)
}
