% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucpoly.R
\name{aucpoly}
\alias{aucpoly}
\title{Draw/color/shade area under a curve.}
\usage{
aucpoly(x, f, a, b, ...)
}
\arguments{
\item{x}{Vector of values at which to evaluate the curve.}

\item{f}{Function for the curve.}

\item{a}{Lower bound of the interval.}

\item{b}{Upper bound of the interval.}

\item{...}{Additional arguments to \code{polygon}.}
}
\description{
This is a utility function mainly for drawing/coloring/shading the area under the curve of a fuction (typically a probability density function).
}
