% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baserun.R
\docType{data}
\name{baserun}
\alias{baserun}
\title{Data from a study of running times to second base.}
\format{A data frame 22 observations and three variables.
\describe{
   \item{round:}{average run time (in seconds) for the round out route}
   \item{narrow:}{average run time (in seconds) for the narrow angle route}
   \item{wide:}{average run time (in seconds) for the wide angle route}
}}
\source{
Woodward, W. F. (1970). A comparison of base running methods in baseball. MSc Thesis, Florida State University.

Hollander, M., Wolfe, D. A., & Chicken, E. (2014). \emph{Nonparametric statistical methods} (3rd edition). Hoboken, New Jersey: Wiley.
}
\usage{
baserun
}
\description{
Each of 22 baseball players ran from home plate to second base six times, twice for each of three routes. The times recorded here are the average running time for the two runs for each route between a point 35 feet from home plate to 15 feet short of second base.
}
\note{
The data are originally from Woodward (1970) but are also discussed in Hollander, Wolfe, and Chicken (2014).
}
\keyword{datasets}
