% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beall.R
\docType{data}
\name{beall}
\alias{beall}
\title{Insecticide randomized block experiments.}
\format{A data frame 550 observations and five variables:
\describe{
   \item{treatment:}{treatment conditions (see below)}
   \item{experiment:}{experiment number and table from Beall (1942)}
   \item{species:}{insect species counted as Pyrausta nubilalis (Pn), Leptinotarsa decemlineata (Ld), or Phlegethontius quinquemaculata (Pq)}
   \item{block:}{block of plots}
   \item{count:}{number of insects}
}}
\source{
Beall, G. (1939). Methods for estimating the population of insects in a field. \emph{Biometrika}, \emph{30}, 422-439.

Beall, G., Stirrett, G. M., & Conners, L. L. (1939). A field experiment on the control of the European corn borer, \emph{Pyrausta nubilalis} Hubn., by \emph{Beauveria Bassiana} Vuill. II. \emph{Scientific Agriculture}, \emph{19}, 531-534.

Beall, G. (1942). The transformation of data from entomological field experiments so that the analysis of variance becomes applicable. \emph{Biometrika}, \emph{32}, 243-262.
}
\usage{
beall
}
\description{
Data from several studies using randomized block designs to investigate the effect of several fungal insecticides.
}
\note{
The data are from Beall (1942) but the data and methods are also described in Beall (1939) and Beall, Stirrett, and Conners (1939). The counts were obtained using random sampling within plots. Beall et al. (1939) describe the treatments for the first two experiments as different amounts of grams of fungal spores per acre (10, 40, 10, 40, 10, 40, and 0, respectively) and date of application (July 9, July 9, July 15, July 15, July 20, July 20, and none, respectively). Furthermore the data from the first two experiments are observations from two different time points as described in Beall et al. (1939). Beall (1942) implies that a similar structure holds for the experiments for the other two species, although it is not clear if the treatments were the same for these experiments. The data from experiment 7 also appear in \code{\link[datasets]{InsectSprays}} although without the block information.
}
\seealso{
\code{\link[datasets]{InsectSprays}}
}
\keyword{datasets}
