% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bliss.R
\docType{data}
\name{bliss}
\alias{bliss}
\title{Mortality of confused flour beetles from carbon disulphide.}
\format{
A data frame with 16 observations and three variables:
\describe{
   \item{concentration:}{concentration of gaseous carbon disulphide (mg/l)}
   \item{dead:}{number of dead beetles after five hours of exposure}
   \item{exposed:}{number of beetles exposed}
}
}
\source{
Bliss, C. I. (1935). The calculation of the dosage-mortality curve. \emph{Annals of Applied Biology}, \emph{22}, 134-167. 

Strand, A. L. (1930). Measuring the toxicity of insect fumigants. \emph{Industrial and Engineering Chemistry: Analytical Edition}, \emph{2}, 4-8.
}
\usage{
bliss
}
\description{
Data from an experiment investigating the effect of gaseous carbon disulphide on the mortality of confused flour beetles (\emph{Tribolium confusum}). Cloth cages of batches of approximately thirty beetles were suspended in a flask above a fixed volume of liquid carbon disulphide. The number of dead beetles after five hours of exposure was recorded.
}
\note{
These data are frequently featured in textbooks and R packages, but usually the two replicates for each concentration are aggregated. The data as shown here were featured in Bliss (1935) but are originally from Strand (1930).
}
\keyword{datasets}
