% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bole.R
\docType{data}
\name{bole}
\alias{bole}
\title{Tree bole volume fixed area plot survey.}
\format{
A data frame with 60 observations and two variables:
\describe{
   \item{volume:}{total plot tree bole volume in cubic meters}
   \item{stratum:}{sampling stratum (softwood, mixed, or hardwood stratum)}
}
}
\source{
Cunia, T. (1979). \emph{Basic designs for survey sampling: Simple stratified, cluster, and systematic sampling}. Forest Biometry Monograph Series 3, College of Environmental Science and Forestry, State University of New York, Syracuse, New York. 

Gregoire, T. G. & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
bole
}
\description{
Data from a stratified random sample of 0.1ha circular plots where total standing tree bole volume was recorded for the trees within each plot. The total area of the softwood, mixed, and hardwood strata are 42541ha, 17250ha, and 32435ha, respectively.
}
\keyword{datasets}
