% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceriodaphnia.R
\docType{data}
\name{ceriodaphnia}
\alias{ceriodaphnia}
\title{Ceriodaphnia toxicity study.}
\format{A data frame with 50 observations and two variables:
\describe{
   \item{concentration}{percent concentration of effluent}
   \item{count}{number of young produced}
}}
\source{
Bailer, A. J. & Oris, J. T. (1997). Estimating inhibition concentrations for differing response scales using generalized linear models. \emph{Environmental Toxicology and Chemistry}, \emph{16}, 1554-1559.
}
\usage{
ceriodaphnia
}
\description{
Data from a study of the number of young produced by Ceriodaphnia dubia exposed for seven days to varying concentrations of effluent.
}
\keyword{datasets}
