% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceriodaphniastrain.R
\docType{data}
\name{ceriodaphniastrain}
\alias{ceriodaphniastrain}
\title{Toxicity study with two strains of ceriodaphnia.}
\format{A data frame with 70 observations and three variables:
\describe{
   \item{count}{number of organisms}
   \item{concentration}{concentration of a component of jet engine fuel}
   \item{strain}{strain of \emph{Ceriodaphnia}}
}}
\source{
Myers, R. H., Montgomery, D. C., Vining, G. G., & Robinson, T. J. (2010). \emph{Generalized linear models with applications in engineering and the sciences}. Hoboken, New Jersey: Wiley.
}
\usage{
ceriodaphniastrain
}
\description{
Data from a study of the effect of varying concentrations of a component of jet engine fuel on reproduction of two strains \emph{Ceriodaphnia}.
}
\note{
No original source for the data is given so they are presumed to be simulated.
}
\keyword{datasets}
