% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlamydia.R
\docType{data}
\name{chlamydia}
\alias{chlamydia}
\title{Screening tests for \emph{Chlamydia trachomatis}.}
\format{A data frame of 4583 observations and six variables.
\describe{
   \item{count:}{number of samples}
   \item{Syva-DFA:}{test result from Syva-DFA test}
   \item{Syva-EIA:}{test result from Syva-EIA test}
   \item{Abbott-EIA:}{test result from Abott-EIA test}
   \item{GenProbe:}{test result from GenProbe test}
   \item{Sanofi-EIA:}{test result from Sanofi-EIA test}
   \item{culture:}{test result from cell culture}
}}
\source{
Hadgu, A. & Qu, Y. (1998). A biomedical application of latent class models with random effects. \emph{Applied Statistics}, \emph{47}, 603-616.

Lindsey, J. K. (1999). \emph{Models for repeated measurements} (2nd Edition). New York: Oxford.
}
\usage{
chlamydia
}
\description{
Results of six tests for \emph{Chlamydia trachomatis} using samples from 4583 women.
}
\note{
Note that the data are in aggregated form showing the number of samples that resulted in a particular profile of test results. All test results are binary indicator variables for positive test results. The data are from Hadgu and Qu (1998) but are also featured in Lindsey (1999) who suggests a latent class model with or without a continuous random effect, and with or without treating the cell culture as a gold standard.
}
\keyword{datasets}
