% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Inferences for contrasts for linear models (experimental).}
\usage{
contrast(model, a, b, a2, b2, fcov = vcov, level = 0.95, cnames = NULL,
  ...)
}
\arguments{
\item{model}{Model object. Currently only objects of class \code{lm} are accepted.}

\item{a}{List or data frame defining a linear combination.}

\item{b}{List or data frame defining a linear combination.}

\item{a2}{List or data frame defining a linear combination.}

\item{b2}{List or data frame defining a linear combination.}

\item{fcov}{Function for estimating the variance-covariance matrix of the model parameters.}

\item{level}{Confidence level in (0,1).}

\item{cnames}{Labels for the contrasts.}

\item{...}{Arguments to pass to \code{fcov}.}
}
\description{
This is a function allows one to obtain standard inferences (i.e., point estimates, standard errors, confidence intervals, etc.) concerning one or more contrasts.
}
\details{
Details go here.
}

