% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cysts.R
\docType{data}
\name{cysts}
\alias{cysts}
\title{Settling time of amoebic cysts in water.}
\format{
A data frame with 21 observations and four variables.
\describe{
   \item{diameter:}{average diameter of cysts in micrometers}
   \item{temp:}{temperature in degrees Celsius}
   \item{time:}{settling time in seconds}
   \item{n:}{number of observations}
}
}
\source{
Chang, S. L. (1945). Sedimentation in water and the specific gravity of cysts of \emph{Entamoeba histolytica}. \emph{American Journal of Epidemiology}, \emph{41}, 156--163.

Rice, J. A. (1998). \emph{Mathematical statistics and data analysis}. Pacific Grove, CA: Wadsworth & Brooks/Cole.
}
\usage{
cysts
}
\description{
Data from an experiment on the settling time of amoebic cysts in water for the purpose of water purification. Observations were made of settling time at three different temperatures. The diameter of the cysts was also recorded. Note that each observation is an average of several observations.
}
\note{
These data are originally from Chang (1945) but are also featured in Rice (2007).
}
\keyword{datasets}
