% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Survival of diabetic men.}
\format{A data frame with four observations and four variables.
\describe{
   \item{age}{less than 40 or 40 years or older}
   \item{insulin}{dependence on insulin (yes or no)}
   \item{dead}{number of deceased individuals}
   \item{censored}{number of individuals with alive or right-censored survival times}
}}
\source{
Gatling, W., Mullee, M., & Hill, R. (1988). General characteristics of a community-based diabetic population. \emph{Practical diabetes}, \emph{5}, 104-107.

Julious, S. A. & Mullee, M. A. (1994). Confounding and Simpson's paradox. \emph{British Medical Journal}, 1480-1481.
}
\usage{
diabetes
}
\description{
Grouped survival data for diabetic men by age and insulin dependence.
}
\note{
@note Julious and Mullee (1994) used these data from the Poole diabetic cohort to illustrate Simpson's paradox which occurs when the relationship between insulin dependence and survival is examined with and without controlling for age. The data are described by Gatling, Mullee, and Hill (1988) although they do not discuss survival.
}
\keyword{datasets}

