% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmethod.R
\name{dmethod}
\alias{dmethod}
\title{Delta method for regression model objects (experimental).}
\source{
Mandel, M. (2013). Simulation-based confidence intervals for functions with complicated derivatives. \emph{The American Statistician}, \emph{62}, 76-81.
}
\usage{
dmethod(object, pfunc, pname, cfunc = coef, vfunc = vcov, tfunc, B = 0,
  level = 0.95, ...)
}
\arguments{
\item{object}{Model object. Just about any object can be specified provided that functions can also be specified to extract the parameter estimates and the estimated covariance matrix of the estimators.}

\item{pfunc}{Character object that is function of the parameters that can be evaluated by R. The function can return a scalar or a vector. Parameters must be referenced by name as specified by \code{pname}.}

\item{pname}{Names of the parameters extracted by \code{cfunc}.}

\item{cfunc}{Function for extracting the parameter estimates from \code{object} (default is \code{coef}).}

\item{vfunc}{Function for extracting the estimated covariance matrix of the estimator from \code{object} (default is \code{vcov}).}

\item{tfunc}{Function for transforming the result of \code{pfunc} for cases when the sampling distribution is thought to be more approximately normal on some other scale (e.g., log).}

\item{B}{Number of bootstrap samples. The default (\code{B = 0}) results in numerical differentiation instead.}

\item{level}{Confidence level in (0,1) (default is 0.95).}

\item{...}{Optional arguments for \code{jacobian}.}
}
\description{
This function applies the "delta method" to functions of parameters of model objects, provided that functions can be specified to extract the parameter estimates and the (estimated) covariance matrix of the estimator from the object. Derivatives are computed using numerical (not symbolic) differentiation. Alternatively standard errors can be approximated using a bootstrap approach described by Mandel (2013).
}
\details{
By default the function applies the delta method using numerical differentiation. However if \code{B} > 0 then a bootstrap method described by Mandel (2013) which avoids needing to compute derivatives.
}
