% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogscats.R
\docType{data}
\name{dogscats}
\alias{dogscats}
\title{Veterinarian's survey of expenses of dogs and cats.}
\format{
A data frame with 50 observations and three variables:
\describe{
   \item{type:}{animal type (dog or cat)}
   \item{visits:}{number of visits}
   \item{expenses:}{total expenses in dollars}
}
}
\source{
Levy, P. S., & Lemeshow, S. (1999). \emph{Sampling of populations: Methods and applications} (3rd Edition). New York: Wiley.
}
\usage{
dogscats
}
\description{
Simple random sample of patient records of a veterinarian. The data can be post-stratified by animal type knowing that in the population of 1300 patients there are 850 dogs and 450 cats.
}
\note{
These data are presumed to be simulated.
}
\keyword{datasets}
