% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinking.R
\docType{data}
\name{drinking}
\alias{drinking}
\title{Hypothetical data for demonstrating Poisson zero-inflation.}
\format{A data frame with 1400 observations and three variables:
\describe{
   \item{sex:}{sex of the respondent (male or female)}
   \item{age:}{age of the respondent (years)}
   \item{drinks:}{number of reported alcoholic drinks in the last week}
}}
\usage{
drinking
}
\description{
Hypothetical data for demonstrating Poisson zero-inflation.
}
\keyword{datasets}
