% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsurvbin.R
\name{dsurvbin}
\alias{dsurvbin}
\title{Create data frame with binary response variables for discrete survival analysis (experimental).}
\usage{
dsurvbin(data, y, event, case.name = "case", time.name = "t",
  resp.name = "y", open = FALSE, reverse = FALSE)
}
\arguments{
\item{data:}{The data frame containing the time variable.}

\item{y:}{Name of the time variable in \code{data}.}

\item{event:}{Indicator variable for observed (i.e., not censored) events where \code{event = 1} if the event was observed at \code{y} and \code{event = 0} if the event had not yet occurred by time \code{y}. If missing then it is assumed that no observations are censored.}

\item{case.name:}{Variable name for cases of binary responses (i.e., all binary responses corresponding to a single survival time).}

\item{time.name:}{Variable name prefix for the time point of each binary response.}

\item{resp.name:}{Variable name for the binary response variables.}

\item{open:}{Logical for whether the observed time point should be considered as corresponding to an open time interval where the right endpoint is infinity (effectively equivalent to assuming that the maximum observed value of \code{y} is right-censored). Default is FALSE.}

\item{reverse:}{Reverse the binary indicator so that \eqn{P(Y_t = 1) = P(T > t|T \ge t)}. Default is FALSE.}
}
\description{
This is a function takes an existing data frame with a discrete time variable and coverts the time variable into a set of binary response variables for modeling discrete survival time.
}
\details{
Assuming survival time is integer-valued as \eqn{T = 1,2,\dots,k}, the probability of a given response can be modeled as \eqn{P(T = 1) = 1 - \lambda(1)} and \deqn{P(T = t) = \lambda(t)(1 - \lambda(j-1))(1 - \lambda(j-2))\dots(1 - \lambda(1))} for \eqn{T > 1}, where \eqn{\lambda(t) = P(T = t|T \ge t)} is the hazard function. If we define a set of binary response variables as \eqn{Y_k = 1} if \eqn{k = t} and \eqn{Y_k = 0} if \eqn{t > k}, then \eqn{P(T = 1) = P(Y_1 = 1)} and \deqn{P(T = t) = P(Y_t = 1)P(Y_{t-1} = 0)P(Y_{t-2} = 0)\dots P(Y_1 = 0)} for \eqn{T > 1}. Thus the likelihood function for \eqn{T} is equivalent to that of the product of \eqn{T} independent binary responses, and so discrete survival time can be modeled as a set of binary response variables using logistic regression or other models for independent binary responses.
}
