% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dungcrabs.R
\docType{data}
\name{dungcrabs}
\alias{dungcrabs}
\title{Change in carapace size of female dungeness crabs after molting.}
\format{A data frame with 472 observations and four variables:
\describe{
   \item{presz}{carapace size (mm) before molting}
   \item{postsz}{carapace size (mm) after molting}
   \item{year}{fishing season for captive crabs}
   \item{source}{source of crab (re-capture or captive)}
}}
\source{
Nolan, D. & Speed, T. P. (2009). \emph{Stat Labs: Mathematical statistics through applications}. New York: Springer. See also the references cited therein.

Hankin, D. G., Diamond, N., Mohr, M. S., & Ianelli, J. (1989). Growth and reproductive dynamics of adult female Dungeness crabs (Cancer magister) in northern California. \emph{Journal du Conseil: ICES Journal of Marine Science}, \emph{46(1)}, 94--108.
}
\usage{
dungcrabs
}
\description{
Data collected from a study of the change in the carapace size of female dungeness crabs (\emph{Metacarcinus magister}) before and after molting. Crabs were either captured prior to molting and then allowed to molt in captivitiy, or were re-captured after being measured and tagged. See Nolan and Speed (2009) and Hankin et al. (1989) for details.
}
\keyword{datasets}

