% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{data}
\name{families}
\alias{families}
\title{Survey of weekly family income and food expenditure.}
\format{A data frame with 33 observations and three variables:
\describe{
   \item{size}{number of family members}
   \item{income}{weekly family income}
   \item{cost}{weekly family expenditure on food}
}}
\source{
Cochran, W. G. (1977). \emph{Sampling techniques} (3rd Edition). New York: Wiley.
}
\usage{
families
}
\description{
These data are from a hypothetical survey of family income and food expenditure. No information
is given about the population.
}
\keyword{datasets}

