% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireants.R
\docType{data}
\name{fireants}
\alias{fireants}
\title{Study of the effect of attractant age on attracting fire ants.}
\format{A data frame with 360 observations and three variables:
\describe{
   \item{day}{day that trap was deployed, and age of attractant}
   \item{count}{number of fire ants trapped}
   \item{group}{traps with attractant (Treatment) or without (Control)}
}}
\source{
These data were obtained from \url{http://people.stat.sfu.ca/~cschwarz/Stat-650/Notes/PDFbigbook-JMP/JMP-part018.pdf}.
}
\usage{
fireants
}
\description{
These data are from a hypothetical study of the effect of the age of a pheromone attractant on attracting fire ants (\emph{Wasmannia auropunctata}). Each day traps with and without the attractant were put in random locations, and 24 hours later the number of ants within each trap was observed. New traps were randomly deployed each day. But the source batch of attractant was one day older each day.
}
\keyword{datasets}

