% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallfly.R
\docType{data}
\name{gallfly}
\alias{gallfly}
\title{Truncated counts of gallfly eggs and gall-cells.}
\format{
A data frame 1411 observations and three variables:
\describe{
   \item{y:}{number of eggs or gall-cells}
   \item{year:}{year of observations (1935 or 1936)}
   \item{stage:}{developmental stage observed (egg or gall)}
}
}
\source{
Finney, D. J. & Varley, G. C. (1955). An example of the truncated Poisson distribution. \emph{Biometrics}, \emph{11}, 387-394.

Varley, G. C. (1947). The natural control of population balance in the knapweed gall-fly (Erophora jaceana). \emph{Journal of Animal Ecology}, \emph{16}, 139-187.
}
\usage{
gallfly
}
\description{
Data from a study that investigated the number of eggs and gall-cells from gall flies (Urophora jaceana) during two seasons in the flower heads of knapweed (Centaurea nemoralis).
}
\note{
These data were used by Finney and Varley (1955) as an example for modeling using a truncated Poisson distribution. For reasons that are not entirely clear, no flower heads with zero eggs or gall-cells were reported. The original source of the data is Varley (1947).
}
\keyword{datasets}
