% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impairment.R
\docType{data}
\name{impairment}
\alias{impairment}
\title{Degree of mental impairment.}
\format{A data frame of 40 observations and three variables:
\describe{
   \item{impairment:}{degree of mental impairment (none, mild, moderate, impaired)}
   \item{ses:}{socioeconomic status (low or high)}
   \item{events:}{life events index score}
}}
\usage{
impairment
}
\description{
Degree of mental impairment as a function of socioeconomic status and life events index.
}
\keyword{datasets}
