% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insulin.R
\docType{data}
\name{insulin}
\alias{insulin}
\title{Insulin assay.}
\format{A data frame with 56 observations and four variables:
\describe{
   \item{y:}{conversion of (3-3H) glucose to toluene-extractable lipids in isolated rat fat cells}
   \item{insulin1:}{dose of standard insulin (pmol/liter)}
   \item{insulin2:}{dose of A1-B29 suberoyl insulin (pmol/liter)}
   \item{ratio:}{ratio of standard insulin to A1-B29 suberoyl insulin}
}}
\source{
Darby, S. C. & Ellis, M. J. (1976). A test of synergism between two drugs. \emph{Applied Statistics}, \emph{25}, 296-299.

McCullagh, P. & Nelder, J. A. (1989). \emph{Generalized linear models} (2nd Edition). Boca Raton: Chapman & Hall/CRC.
}
\usage{
insulin
}
\description{
The data are from a study designed to determine if the effect of two insulins is additive or synergistic.
}
\details{
These data were featured in Darby and Ellis (1976) and also McCullagh and Nelder (1989). In particular see the latter for a discussion of a nonlinear gamma regression model for these data.
}
\keyword{datasets}
