% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironretention.R
\docType{data}
\name{ironretention}
\alias{ironretention}
\title{Iron retention in mice.}
\format{A data frame of 108 observations on three variables:
\describe{
   \item{retention:}{percent of iron administered that was retained}
   \item{concentration:}{concentration of iron administered in millimolars}
   \item{ion:}{type of iron (ferrious or ferric)}
}}
\source{
Rice, J. A. (1998). \emph{Mathematical statistics and data analysis}. Pacific Grove, CA: Wadsworth & Brook/Cole.
}
\usage{
ironretention
}
\description{
Data from a randomized experiment on the retention of ferrous (Fe2+) and ferric (Fe3+) iron administered to mice at concentrations of 10.2, 1.2, and 0.3 millimolars. The percent of iron retained was measured.
}
\keyword{datasets}
