% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larkspur.R
\docType{data}
\name{larkspur}
\alias{larkspur}
\title{Sample of transects of pale white larkspur.}
\format{A data frame with 20 observations and two variables:
\describe{
   \item{plants}{abundance in a sampled transect}
   \item{area}{sampled transect area in square meters}
}}
\source{
Stehman, S. V., & Salzer, D. W. 2000. Estimating density from surveys employing unequal-area belt transects. \emph{Wetlands}, \emph{20}, 512-519.

Gregoire, T. G., & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
larkspur
}
\description{
Sample of 20 one-meter belt transects to estimate the abundance of pale white larkspur in a given region. The total number of sampling units in the sampling frame is 150 belt transects with a total area of 5897 square meters.
}
\note{
Although not stated explicitly in the source, the sample was presumably obtained using simple random sampling.
}
\keyword{datasets}
