% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafarea.R
\docType{data}
\name{leafarea}
\alias{leafarea}
\title{Survey of leaf surface area.}
\format{
A data frame with 20 observations and two variables:
\describe{
   \item{area:}{leaf surface area (cubic cm)}
   \item{lengthwidth:}{leaf length times width (cubic cm)}
}
}
\source{
Gregoire, T. G. & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
leafarea
}
\description{
A simple random sample of 20 leaves from a population of 744 leaves of "shining gum" (\emph{Eucalyptus nitens}). The area of all 744 leaves was approximated by computing length times width, with a total area of 57266.6 cubic cm. The leaf surface area was accurately measured for only this sample of 20 leaves.
}
\note{
The data are from Gregoire and Valentine (2007).
}
\keyword{datasets}
