% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincon.R
\name{lincon}
\alias{lincon}
\alias{lincon.lm}
\alias{lincon.nls}
\alias{lincon.glm}
\alias{lincon.lmerMod}
\alias{lincon.glmerMod}
\title{Linear combinations of regression model parameters.}
\usage{
\method{contrast}{lm}(model, a, b, df, tf, cnames, level = 0.95, fcov = vcov, ...)
\method{contrast}{nls}(model, a, b, df, tf, cnames, level = 0.95, fcov = vcov, ...)
\method{contrast}{glm}(model, a, b, df, tf, cnames, level = 0.95, fcov = vcov, ...)
\method{contrast}{lmerMod}(model, a, b, df, tf, cnames, level = 0.95, fcov = vcov, ...)
\method{contrast}{glmerMod}(model, a, b, df, tf, cnames, level = 0.95, fcov = vcov, ...)
}
\arguments{
\item{model}{Model object. Currently only objects of class \code{lm} and \code{nls} are accepted.}

\item{...}{Not used.}

\item{a}{Vector or matrix defining the \eqn{a_j} coefficients of the linear combination(s). If omitted then this defaults to the identity matrix to provide inferences for each parameter similar to the \code{summary} function.}

\item{b}{A scalar or vector defining the \eqn{b} coefficient(s) of the linear combination. Assumed to be zero if missing.}

\item{df}{Optional degrees of freedom. If left missing the residual degrees of freedom will be used except for GLMs with \code{family = poisson} or \code{family = binomial} in which case an infinite degrees of freedom is used.}

\item{tf}{Optional transformation function to apply to the point estimate(s) and confidence interval limits (e.g., \code{tf = exp} for a logistic model to estimate odds or odds ratios).}

\item{cnames}{Optional vector of contrast names. If left missing the contrast coefficients are shown. If FALSE then no names are shown.}

\item{level}{Confidence level in (0,1). Default is 0.95.}

\item{fcov}{Function for estimating the variance-covariance matrix of the model parameters.}
}
\description{
This is a function allows one to obtain standard inferences (i.e., point estimates, standard errors, confidence intervals, etc.) concerning any linear combination of regression model parameters.
}
\details{
For a regression model with a linear component (e.g., linear and generalized linear models) with parameters \eqn{\beta_0, \beta_1, \dots, \beta_p} a linear combination is defined as \deqn{a_0\beta_0 + a_1\beta_1 + \cdots + a_p\beta_p + b.} For a nonlinear regression model with parameters \eqn{\theta_1, \theta_2, \dots, \theta_q} a linear combination is defined as \deqn{a_1\theta_1 + a_2\theta_2 + \cdots + a_q\theta_q + b.} Inferences for the linear combination are based on either exact (normal theory) or Wald (asymptotic) test statistics and confidence intervals. The estimated standard error(s) of the linear combinations are computed using any specified function for estimating the variance-covariance matrix of the model parameters.
}
\examples{
myreg <- lm(Gas ~ Insul + Temp + Insul:Temp, data = MASS::whiteside)
# same as summary(myreg)
lincon(myreg)
# slope with respect to temperature after insulation
lincon(myreg, a = c(0,0,1,1))
# expected gas consumption before and after insulation at 5 degrees celsius
lincon(myreg, a = matrix(c(1,0,5,0,1,1,5,5), 2, 4, byrow = TRUE), cnames = c("Before","After"))
# change in expected gas consumption at five degrees celsius from adding insulation
lincon(myreg, a = c(0,1,0,5))

}
