% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margeff.R
\name{margeff}
\alias{margeff}
\title{Estimation of marginal effects regression models (experimental).}
\usage{
margeff(model, a, b, df, cnames, delta = 1, level = 0.95, fcov = vcov,
  ...)
}
\arguments{
\item{model}{Model object of class \code{lm}, \code{glm}, or \code{nls}.}

\item{a}{List or data frame defining values of the explanatory variables.}

\item{b}{List or data frame defining values of the explanatory variables.}

\item{df}{Degrees of freedom for the confidence interval. If omitted it is extracted from the model object.}

\item{cnames}{Optional names for the marginal effects.}

\item{delta}{Divisor for the marginal effect (default is one).}

\item{level}{Confidence level in (0,1).}

\item{fcov}{Function for estimating the variance-covariance matrix of the model parameters.}

\item{...}{Arguments to pass to \code{fcov}.}
}
\description{
This function estimates discrete marginal effects (i.e., changes in the expected response) for (generalized) linear and nonlinear regression models. It can also approximate "instantaneous" marginal effects.
}
\details{
A (discrete) marginal effect is defined as \eqn{[E(Y|X = a) - E(Y|X = b)]/\delta} where \eqn{a} and \eqn{b} represent specified values of the explanatory variables. Typically these differ only in terms of the value of one explanatory variable to estimate the marginal effect of changing that explanatory variable when the other explanatory variables are held constant at specified values. For continuous explanatory variables the "instantaneous" marginal effect is a limiting case. For example, if there are two explanatory variables --- \eqn{X_1} and \eqn{X_2} --- the marginal effect of \eqn{X_1} at \eqn{X_1 = x_1} and \eqn{X_2 = x_2} can be defined as the limit of \eqn{[E(Y|X_1 = x_1 + \delta, X_2 = x_2) - E(Y|X_1 = x_1, X_2 = x_2)]/\delta} as \eqn{\delta} goes to zero (i.e., the derivative \eqn{E(Y|X_1=x_1,X_2=x_2)} with respect to \eqn{x_1}). This can be approximated by setting \eqn{\delta} to a sufficiently small number.
}
\examples{
m <- glm(cbind(deaths, total - deaths) ~ insecticide * deposit, 
 family = binomial, data = insecticide) 
# discrete marginal effect of increasing deposit from 4 to 6 for each insecticide  
margeff(m, 
 a = list(deposit = 3, insectide = levels(insecticide$insecticide)),
 b = list(deposit = 2, insectide = levels(insecticide$insecticide)),
 cnames = levels(insecticide$insecticide))
# approximate "instantaneous" marginal effects for each insecticide at a deposit of 5 
margeff(m, 
 a = list(deposit = 5 + 0.001, insectide = levels(insecticide$insecticide)),
 b = list(deposit = 5, insectide = levels(insecticide$insecticide)),
 cnames = levels(insecticide$insecticide), delta = 0.001)
}
