% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothprogeny.R
\docType{data}
\name{mothprogeny}
\alias{mothprogeny}
\title{Sex ratios in moth progeny.}
\format{A data frame 60 observations and three variables:
\describe{
   \item{male:}{number of male offspring}
   \item{female:}{number of female offspring}
   \item{locality:}{location the moth/progeny was sampled}
}}
\source{
Lewis, T. (1987). Uneven sex ratios in the light-brown apple moth: A problem in outlier detection. In D. J. Hand & B. S. Everitt (Eds.), \emph{The statistical consultant in action} (121-133). Cambridge: Cambridge University Press.
}
\usage{
mothprogeny
}
\description{
Data from a study of the progeny of light-brown apple moths (\emph{Epiphyas postvittana}) in four localities in Australia. Each observation is the progeny of a single adult female moth.
}
\note{
The appearance of several progeny with few or no male offspring is thought to be due to a particular genetic condition, so that the sex ratios may be due to a finite or continuous mixture. See Lewis (1987) for details about the study and some statistical issues.
}
\keyword{datasets}
