% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nabr.R
\docType{data}
\name{nabr}
\alias{nabr}
\title{Toxicity of sodium bromide on the reproduction of Daphnia magna.}
\format{A data frame of 50 observations on two variables:
\describe{
   \item{young:}{number of offspring per adult over 23 days}
   \item{concentration:}{concentration of sodium bromide (micrograms per liter)}
}}
\source{
OMaul, A., El-Shaarawi, A. H., and F\'{e}rard, J. F. (1991). Application of negative binomial regression models to the analysis of quantal bioassay data. \emph{Environmetrics}, \emph{2}, 253--261. The data are also featured in Maul, A. (1992). Application of generalized linear models to the analysis of toxicity test data. \emph{Environmental Monitoring and Assessment}, \emph{23}, 153--163.
}
\usage{
nabr
}
\description{
The data are from study that exposed populations of Daphnia magna to different concentrations of sodium bromide (NaBr) over 23 days. The number of offspring produced per adult was observed.
}
\keyword{datasets}
