% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naevi.R
\docType{data}
\name{naevi}
\alias{naevi}
\title{Case-control study of children with spina bifida.}
\format{A data frame of 42 observations and four variables:
\describe{
   \item{sex:}{child's sex (male or female)}
   \item{age:}{child's age in years}
   \item{case:}{number of naevi for the case child}
   \item{control:}{number of naevi for the control child}
}}
\source{
Pawitan, Y. (2013). \emph{In all likelihood: Statistical modeling and inference using likelihood}. Oxford University Press.
}
\usage{
naevi
}
\description{
Data from a matched case-control study examining the association between spina bifida and the incidence of naevi (dark raised skin lesions). Cases were matched 1:1 with controls using sex and age. Cases are children with spina bifida whereas controls are children who do not have it.
}
\note{
The data are from an example from Pawitan (2013). The original source is unknown.
}
\keyword{datasets}
