% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nectriacanker.R
\docType{data}
\name{nectriacanker}
\alias{nectriacanker}
\title{Fixed area sample of tree diameter and abundance of nectria cankers.}
\format{A data frame with 38 observations and three variables:
\describe{
   \item{plot}{fixed area plot number}
   \item{dbh}{tree diameter at breast height in cm}
   \item{cankers}{number of observed cankers on tree}
}}
\source{
Barrett, J. P. & Nutt, M. E. (1979). \emph{Survey sampling in the environmental sciences: A computer approach}. Wentworth, NH: COMPress, Inc.

Gregoire, T. G. & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
nectriacanker
}
\description{
These data are from a sample of fixed area plots, each of area 0.01ha from a 10ha tract of land. Tree diameter and the number of cankers were observed for each tree in each plot.
}
\keyword{datasets}
