% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralgia.R
\docType{data}
\name{neuralgia}
\alias{neuralgia}
\title{Treatment of post-herpetic neuralgia.}
\format{A data frame of 18 observations and five variables.
\describe{
   \item{pain:}{reported pain cessation at six weeks}
   \item{treatment:}{treatment condition (vincristine or saline)}
   \item{age:}{patient age (in years) at treatment}
   \item{sex:}{patient sex}
   \item{complaint:}{duration (months) of pre-treatment complaint} 
}}
\source{
Layman, P. R., Argyras, E., & Glynn, C. J. (1986). Iontophoresis of vincristine versus saline in post-herpetic neuralgia. A controlled trial. \emph{Pain}, \emph{25}, 165-170.

Piegorsch, W. W. (1992). Complementary log-log regression for generalized linear models. \emph{The American Statistician}, \emph{46}, 94-99.
}
\usage{
neuralgia
}
\description{
These data are from a study in which patients with post-herpetic neuralgia were randomly assigned to receive iontophoresis of either vincristine or saline. The patients were interviewed six weeks after treatment to determine if there was any improvement in the neuralgia.
}
\note{
The data are from Layman, Agyras, and Glynn (1986) and were featured in Piegorsch (1992) as an example of data for a generalized linear model with a complementary log-log link function since time to pain cessation was thought to follow an approximate exponential distribution.
}
\keyword{datasets}
