% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsint.R
\name{nlsint}
\alias{nlsint}
\title{Confidence and prediction intervals for the expected/future response of a nonlinear regression model.}
\usage{
nlsint(object, newdata = eval(object$call$data), interval = c("confidence",
  "prediction"), level = 0.95)
}
\arguments{
\item{object}{An object of class \code{nls} (i.e., an estimated nonlinear regression model).}

\item{newdata}{Data frame of variables at which to compute confidence/prediction intervals.
Defaults to the same data frame used to create \code{object}.}

\item{interval}{Type of interval (confidence or prediction).}

\item{level}{Confidence/tolerance level.}
}
\value{
A data frame of the predicted response, the standard error or prediction error, and the lower
and upper bound of the confidence/prediction interval at each set of variable values in \code{newdata}.
}
\description{
\code{nlsint} computes approximate confidence intervals for the expected response and prediction intervals
for the future response of a nonlinear regression model of class \code{nls}.
}
\details{
Standard errors and the estimated expected response are obtained using the delta method based
on the numerical derivatives and the estimated covariance matrix of the model parameters. The prediction 
use the estimated error variance and implicitly assumes identically distributed normal errors.
}
\examples{
model <- nls(rate ~ Vm * conc / (K + conc), start = c(Vm = 200, K = 0.1), data = subset(Puromycin, state == "treated"))
data <- data.frame(conc = seq(0.02, 1.10, by = 0.01))
nlsint(model, newdata = data)
nlsint(model, newdata = data, interval = "prediction") 
}

