% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photocounts.R
\docType{data}
\name{photocounts}
\alias{photocounts}
\title{Aerial photograph and field counts of dead trees.}
\format{A data frame with 80 observations and two variables:
\describe{
   \item{photo}{count of dead trees from aerial photograph}
   \item{field}{count of dead trees from the field}
}}
\source{
Barrett, J. P. & Nutt, M. E. (1979). \emph{Survey sampling in the environmental sciences: A computer approach}. Wentworth, NH: COMPress, Inc.

Gregoire, T. G. & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
photocounts
}
\description{
Counts of dead trees using aerial photographs and field counts in a 80ha parcel of land divided into 80 units. Field counts are more costly but also more accurate than those based on aerial photographs.
}
\note{
These data can be treated as a census of the land parcel.
}
\keyword{datasets}

