% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pike.R
\docType{data}
\name{pike}
\alias{pike}
\title{Creel surveys of Northern Pike.}
\format{A data frame of 16 observations and two variables:
\describe{
   \item{catch:}{number of fish caught}
   \item{tagged:}{number of tagged fish caught}
}}
\source{
Jessen, R. J. (1978). \emph{Statistical survey techniques}. Wiley.
}
\usage{
pike
}
\description{
Data from a survey of Northern Pike (\emph{Esox lucius}) at Ventura Marsh, Iowa, in October of 1953. A total of 1146 tagged pike were first released into the system, and then 16 hauls were conducted (presumably sampling with replacement).
}
\note{
The data are from Jessen (1978). Jessen states that the data are from R. J. Muncy and R. L. Ridenhour. Note that the summary statistics for catch reported by Jessen do not quite match the data he reported.
}
\keyword{datasets}
