% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pock.R
\docType{data}
\name{pock}
\alias{pock}
\title{Pock lesions on chicken embryos.}
\format{
A data frame of 58 observations on two variables.
\describe{
   \item{Dilution:}{dilution of the viral medium}
   \item{Count:}{number of lesions}
}
}
\source{
Breslow, N. E. (1990). Further studies in the variability of pock counts. \emph{Statistics in Medicine}, \emph{9}, 615--626.
}
\usage{
pock
}
\description{
Data from a study of the number of lesions from viruses of the pox group at several dilutions of the viral medium.
}
\keyword{datasets}
