% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radios.R
\docType{data}
\name{radios}
\alias{radios}
\title{Survey of number of radios in households.}
\format{A data frame of 240 observations and three variables:
\describe{ 
   \item{stratum:}{identifier for the stratum (1 through 240)}
   \item{households:}{number of households in the sampling unit}
   \item{radios:}{number of radios owned by the households in the sampling unit}
}}
\source{
Jessen, R. J. (1978). \emph{Statistical survey techniques}. Wiley.
}
\usage{
radios
}
\description{
Data from a survey of households in the "open country" of Iowa in 1951. A total of 71040 area sampling units were stratified into 120 strata with an equal number of geographically contiguous units per strata (592 units per strata). Two units were sampled from each stratum.
}
\note{
The data are from Jessen (1978).
}
\keyword{datasets}
