% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radonbase.R
\docType{data}
\name{radonbase}
\alias{radonbase}
\title{Survey of radon concentration in homes in Blue Ridge, VA.}
\format{A data frame with 140 observations and two variables:
\describe{
   \item{radon}{ground floor radon concentration in pCi/L}
   \item{basement}{does the house have a basement?}
}}
\source{
Gregoire, T. G. & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
radonbase
}
\description{
Data from a survey of homes in Blue Ridge, VA, from a population of 13,895 homes. The sampling design was simple random sampling. The radon concentration was measured in the ground floor of each home. It was also observed whether or not the home had a basement with concrete walls.
}
\keyword{datasets}

