% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radonminn.R
\docType{data}
\name{radonminn}
\alias{radonminn}
\title{Survey of radon concentration in homes in Minnesota.}
\format{A data frame with 1003 observations and four variables:
\describe{
   \item{county}{county name}
   \item{n}{number of homes in the county and in the sample}
   \item{N}{number of homes in the county}
   \item{radon}{radon concentration in pCi/L}
}}
\source{
Nolan, D. & Speed, T. P. (2009). \emph{Stat Labs: Mathematical statistics through applications}. New York: Springer.

These data were obtained from \url{http://www.stat.ncsu.edu/people/boos/courses/st715/schedule.715.html}.
}
\usage{
radonminn
}
\description{
Data from a survey of radon levels in homes in Minnesota. The survey used a
stratified random sampling design with counties as strata. See Nolan and
Speed (2009) for details about the sampling design.
}
\note{
As noted by Nolan and Speed (2009) there was some non-response but
no information is given for non-responders.
}
\keyword{datasets}

