% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratdiet.R
\docType{data}
\name{ratdiet}
\alias{ratdiet}
\title{Effect of a restricted diet on lifespan of rats.}
\format{
A data frame with 195 observations and two variables:
\describe{
   \item{days:}{lifespan (days)}
   \item{restricted:}{diet restricted (yes or no)}
}
}
\source{
Berger, R. L., Boos, D. D., & Guess, F. M. (1988). Tests and confidence sets for comparing two mean residual life functions. \emph{Biometrics}, \emph{44}, 103--115. 

Yu, B. P., Masoro, E. J., Murata, I., Bertrand, H. A., & Lynd, F. T. (1982). Lifespan study of SPF Fisher 344 male rats fed ad libitum or restricted diets: Longevity, growth, and lean body mass and disease. \emph{Journal of Gerontology}, \emph{37}, 130-141.
}
\usage{
ratdiet
}
\description{
Data from an experiment investigating the effect of a restricted versus an unrestricted diet on the lifespan of rats.
}
\keyword{datasets}
