% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh.R
\docType{data}
\name{rayleigh}
\alias{rayleigh}
\title{Lord Rayleigh's measurements of nitrogen gas.}
\format{A data frame with 15 observations and 2 variables:
\describe{
   \item{mass:}{mass of gas sample in grams}
   \item{method:}{method of isolating the nitrogen gas}
}}
\source{
Larsen, R. D. (1990). Lessons learned from Lord Rayleigh on the importance of data analysis. \emdph{Journal of Chemical Education}, \emph{67(11)}, 925--928.
}
\usage{
rayleigh
}
\description{
Lord Rayleigh (John William Strutt) conducted several experiments in which he tried to isolate samples of nitrogen gas at a fixed volume, temperature, and pressure by removing oxygen from dry air or through the decomposition of nitrogen compounds.
}
\note{
The difference in the mass of nitrogen isolated from atmospheric air (using the "hot iron" or "ferrous hydrate" methods) with nitrogen obtained from chemical decompositions (from nitric oxide, nitrous oxide, or ammonium nitrate) is assumed to be due to the presence of other elements or compounds in the atmospheric air, namely argon.
}
\keyword{datasets}
