% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoak.R
\docType{data}
\name{redoak}
\alias{redoak}
\title{Measurements of red oaks.}
\format{
A data frame with 236 observations and three variables:
\describe{
   \item{dbh:}{diameter at breast height (in)}
   \item{volume:}{tree volume (cubic ft)}
   \item{height:}{tree height (ft)}
}
}
\source{
Beers, T. W. & Gingrich, S. F. (1958). Construction of cubic-foot volume tables for red oak in Pennsylvania. \emph{Journal of Forestry}, \emph{56}, 210--214.

Gregoire, T. G. & Valentine, H. T. (2007). \emph{Sampling strategies for natural resources and the environment}. Boca Raton, FL: Chapman & Hall/CRC.
}
\usage{
redoak
}
\description{
Three measurements of red oaks (\emph{Quercus rubra}): diameter at breast height, volume, and heigh.
}
\note{
The data are from Gregoire and Valentine (2007). They obtained the data from Beers and Gingrich (1958).
}
\keyword{datasets}
