% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowgauge.R
\docType{data}
\name{snowgauge}
\alias{snowgauge}
\title{Calibration study of a snow gauge.}
\format{A data frame with 90 observations and two variables:
\describe{
   \item{density}{density (grams per cubic centimeter)}
   \item{gain}{amplified measurement of gamma photon count}
}}
\source{
Nolan, D. & Speed, T. P. (2009). \emph{Stat Labs: Mathematical statistics through applications}. New York: Springer. See also the references cited therein.
}
\usage{
snowgauge
}
\description{
Data collected from snow gauge that measures snow density using detected gamma photon counts from a radioactive source. Observations of the "gain" (an amplified count) at several known densities of polyethylene blocks to simulate snow were observed.
}
\keyword{datasets}

