% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sole.R
\docType{data}
\name{sole}
\alias{sole}
\title{Data from an experiment of the effect of salinity and temperature on hatching of sole.}
\format{
A data frame 72 observations on four variables.
\describe{
   \item{salinity:}{water salinity in g/kg}
   \item{temp:}{water temperature in degrees Celsius}
   \item{hatched:}{number of eggs that hatched}
   \item{total:}{number of eggs in tank}
}
}
\source{
Alderdice, D. F., & Forrester, C. R. (1968). Some effects of salinity and temperature on early development and survival of the English sole (\emph{Parophyrys vetulus}). \emph{Journal of Fisheries Research Board of Canada}, \emph{25(3)}, 495--521.

Lindsey, J. K., Alderdice, D. F., & Pienaar, L. V. (1970). Analysis of nonlinear models -- the nonlinear response model. \emph{Journal of the Fisheries Board of Canada}, \emph{27}, 765--791.

Lindsey, J. K. (1975). Likelihood analysis and tests for binary data. \emph{Journal of the Royal Statistical Society: Series C}, \emph{24(1)}, 1--16.
}
\usage{
sole
}
\description{
Tanks of eggs of English sole (\emph{Parophrys vetulus}) were exposed to different levels of salinity and temperature. The number of eggs in each tank that hatched was counted.
}
\note{
The data are originally from Woodward Alderdice and Forrester (1968), and are featured in Lindsey, Alderdice, and Pienaar (1970) and Lindsey (1975).
}
\keyword{datasets}
