% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmers.R
\docType{data}
\name{swimmers}
\alias{swimmers}
\title{Ear infections in swimmers.}
\format{A data frame with 287 observations and five variables:
\describe{
   \item{swimmer:}{type of swimmer (Occasional or Frequent)}
   \item{location:}{location of swimming (NonBeach or Beach)}
   \item{age:}{swimmer's age (years)}
   \item{sex:}{swimmer's sex (Female or Male)}
   \item{infections:}{number of self-diagnosed ear infections}
}}
\source{
Hand, D. J., Daly, F., Lunn, A. D., McConway, K. J., Ostrowski, E. (1994). \emph{A handbook of small data sets}. London: Chapman & Hall. The data were downloaded from http://www.statsci.org/data/oz/earinf.html.
}
\usage{
swimmers
}
\description{
Data from an observational study of the frequency of ear infections in swimmers.
}
\keyword{datasets}
