% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\docType{data}
\name{synchrony}
\alias{synchrony}
\title{Infant attention to speech in and out of synchrony.}
\format{A data frame 12 observations and three variables:
\describe{
   \item{subject:}{infant identifier}
   \item{insync:}{percent of time attending to the person speaking in the in-synchrony condition}
   \item{outsync:}{percent of time attending to the person speaking in the out-of-synchrony condition}
}}
\source{
Dodd, B. (1979). Lip reading in infants: Attention to speech presented in- and out-of-synchrony. \emph{Cognitive Psychology}, \emph{11}, 478-484.

Siegel, S. & Castellan, N. J. (1988). \emph{Nonparametric statistics for the behavioral sciences} (Second Edition). New York: McGraw-Hill.
}
\usage{
synchrony
}
\description{
Data from a study that investigated if 10- to 16-week old infants are aware of the synchrony between lip movements and speech sounds. Infants were placed in a soundproof room with a window through which they could see a person speaking. The sound of the speech (delivered through a microphone and speaker in the room with the infant) was either synchronous or out of synchrony due to a 400-millisecond delay. Each infants was observed in both conditions, and the percent of the time they attended to the person speaking was recorded.
}
\note{
The data are from Dodd (1979) and were featured in Siegel and Castellan (1988).
}
\keyword{datasets}
