% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timss.R
\docType{data}
\name{timss}
\alias{timss}
\title{Third International Mathematics and Science Study.}
\format{A data frame with 2427 observations on 10 variables:
\describe{
   \item{region}{geographic region of the U.S. (northeast, south, central, or west)}
   \item{school}{integer identifier of school}
   \item{student}{integer identifier of student}
   \item{sex}{sex of student (female or male)}
   \item{language}{how often the language of the test is spoken at home (always, sometimes, or never)}
   \item{math}{mathematics test score}
   \item{ethnicity}{ethnic category of student (white non-hispanic, black, hispanic, asian, native american, or other)}
   \item{science}{science test score}
   \item{community}{type of locatin of school (village or rural area, outskirts of town or city, close to center of a town or city}
   \item{enrollment}{number of students enrolled at the school}
}}
\source{
TIMSS International Study Center (1999). \emph{Third International Mathematics and Science Study: Technical Report, Volume 1 Design and Development}. Chestnut Hill, MA: Boston College. 

Valliant, R., Dorfman, A. H., & Royall, R. M. (2000). \emph{Finite population sampling and inference: A prediction approach}. New York: Wiley.
}
\usage{
timss
}
\description{
These data are from the Third International Mathematics and Science Study that are part of the Trends in International Mathematics and Science Study (TIMSS) conducted by the National Center for Educational Statistics.
}
\note{
These data were obtained from \link{http://www.wiley.com/WileyCDA/WileyTitle/productCd-0471293415.html}. They are featured in Valliant, Dorfman, and Royall (2000). According to Vallian et al. (2000) these data are from third grade students in schools in the United States. For more information about the data and design see TIMSS (1996).
}
\keyword{datasets}
