% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watercolumn.R
\docType{data}
\name{watercolumn}
\alias{watercolumn}
\title{Phosphorus concentrations in water samples from Willow Creek Reservoir, Oregon.}
\format{
A data frame with 544 observations and five variables:
\describe{
   \item{date}{day the sample was obtained}
   \item{site}{site where the sample was obtained}
   \item{depth}{depth (in meters) at which the sample was obtained}
   \item{tp}{total phosphorus concentration in micrograms per liter}
   \item{dp}{dissolved phosphorus concentration in micorgrams per liter}
}
}
\source{
Burnet, S. H. (2016). \emph{Comparison of two methods to estimate within-lake regenerated nutrients and determine their role in harmful algal blooms} (MS Thesis). University of Idaho.
}
\usage{
watercolumn
}
\description{
These data are the total and dissolved phosphorus concentrations in water column samples from Willow Creek Reservoir in Heppner, Oregon. Each observation of concentration is the mean concentration of three replicate samples.
}
\keyword{datasets}
