% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightheight.R
\docType{data}
\name{weightheight}
\alias{weightheight}
\title{Weight/height ratio and ages of boys.}
\format{A data frame with 72 observations and two variables:
\describe{
   \item{ratio:}{weight/height ratio (lbs/in)}
   \item{age:}{age (months)}
}}
\source{
Eppright, E. S., Fox, H. M., Fryer, B. A., Lamkin, G. H., Vivian, V. M., & Fuller, E. S. (1972). Nutrition of infants and preschool children in the north central region of the United States of America. \emph{World Review of Nutrition and Dietetics}, \emph{14}, 269--332.

Gallant, A. R. & Fuller, W. A. (1973). Fitting segmented polynomial regression models whose joint points have to be estimated. \emph{Journal of the American Statistical Association}, \emph{68}, 144--147.
}
\usage{
weightheight
}
\description{
Data from an observational study of the relationship between the weight/height ratio and ages of boys.
}
\note{
These data are originally from Eppright et al. (1972) and are featured in Gallant and Fuller (1973).
}
\keyword{datasets}
