% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmint.R
\name{glmint}
\alias{glmint}
\title{Confidence intervals for the estimated expected response of generalized linear models (experimental).}
\usage{
glmint(object, newdata, df, level = 0.95)
}
\arguments{
\item{object}{Model object of class \code{glm}.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted the data frame used in creating the model object is used.}

\item{df}{Degrees of freedom for the confidence interval. If omitted it is extracted from the model object.}

\item{level}{Confidence level in (0,1).}
}
\description{
This function computes the estimated expected response and the confidence interval for the expected response for (generalized) linear models.
}
\details{
This function simplifies the process of computing the endpoints of the confidence interval for the expected response on the scale of the response variable for generalized linear models using the \code{predict} function with \code{se.fit = TRUE}. It computes the Wald confidence interval on the scale of the linear predictor and then uses the inverse of the link function to map the interval to the scale of the response variable.
}
\examples{
m <- glm(cbind(y,total-y) ~ density + species, family = quasibinomial, data = rotifer)
d <- expand.grid(density = seq(1.01, 1.07, by = 0.01), species = c("kc","pm"))
glmint(m, newdata = d)

}
