% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincon.R
\name{lincon}
\alias{lincon}
\title{Linear combinations of regression model parameters.}
\usage{
lincon(model, a, level = 0.95, fcov = vcov, ...)
}
\arguments{
\item{model}{Model object. Currently only objects of class \code{lm} are accepted.}

\item{a}{Vector or matrix defining the linear combination(s).}

\item{level}{Confidence level in (0,1).}

\item{fcov}{Function for estimating the variance-covariance matrix of the model parameters.}

\item{...}{Arguments to pass to \code{fcov}.}
}
\description{
This is a function allows one to obtain standard inferences (i.e., point estimates, standard errors, confidence intervals, etc.) concerning any linear combination of regression model parameters.
}
\details{
For a regression model with parameters \eqn{\beta_0, \beta_1, \dots, \beta_p} a linear combination is defined as \eqn{a_0\beta_0 + a_1\beta_1 + \cdots + a_p\beta_p}. Inferences are based on either exact or Wald test statistics and confidence intervals. The estimated standard error(s) of the linear combinations are computed using any specified function for estimating the variance-covariance matrix of the model parameters.
}

