% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnr.R
\name{logitnr}
\alias{logitnr}
\title{Logit link with a natural response parameter.}
\usage{
logitnr(nrp = 0)
}
\arguments{
\item{nrp}{Natural response parameter in [0,1).}
}
\description{
Link function for logistic regression models with a (fixed/known) natural response parameter.
}
\details{
The logit link function with a natural response parameter is defined as \eqn{\log((E(Y_i) - \pi_i)/(1 - E(Y_i))}, where \eqn{Y_i} is the response variable (i.e., a proportion in the case of a binomial distribution) and \eqn{0 \le \pi_i < 1} is the natural response parameter. Note that the usual logit link function is a special case where \eqn{\pi_i = 0}.
}
\note{
User-specified starting values for the regression parameters may be required in cases where \eqn{\pi_i} is substantially larger than zero. One strategy is to obtain these from a model with a smaller value \eqn{\pi_i}.
}
\examples{
# obtain starting values with the natural respone parameter set to 0
tmp <- glm(cbind(total-y, y) ~ density + species, family = binomial(link = logitnr(0)), data = rotifer)
# use starting values with a fixed natural response parameter of 0.1
glm(cbind(total-y, y) ~ density + species, family = binoimal(link = logitnr(0.1), data = rotifer, start = coef(tmp)))
}

