% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglog.R
\docType{data}
\name{loglog}
\alias{loglog}
\title{Log-log link function for generalized linear models.}
\format{An object of class \code{link-glm} of length 6.}
\usage{
loglog
}
\description{
This is the log-log link function for generalized linear models, usually with a binomially-distributed response variable.
}
\details{
The log-log link function is defined as \eqn{\log[-\log[E(Y_i)]]}, where \eqn{Y_i} is the response variable (i.e., a proportion in the case of a binomial distribution). Note that using the log-log link function for the proportion of ``successes'' is the same as using the complementary log-log link function for the proportion of ``failures'' in a generalized linear model. The coefficients from the two models only differ by sign.
}
\examples{
# the following produce the same results up to a reversal of the sign of the coefficients
model.loglog <- glm(cbind(y, total-y) ~ density + species, 
 family = binomial(link = loglog), data = rotifer)
model.cloglog <- glm(cbind(total-y, y) ~ density + species, 
 family = binomial(link = cloglog), data = rotifer)
}
\keyword{datasets}
