% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsint.R
\name{nlsint}
\alias{nlsint}
\title{Confidence and prediction intervals for the (expected) response of a nonlinear regression model (experimental).}
\usage{
nlsint(object, newdata = eval(object$call$data), interval = c("confidence",
  "prediction"), fcov = vcov, level = 0.95, ...)
}
\arguments{
\item{object}{Model object of class \code{nls}.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted the data frame used in creating the model object is used.}

\item{interval}{Type of interval calculation (confidence or prediction).}

\item{fcov}{Function for estimating the variance-covariance matrix of the model parameters.}

\item{level}{Confidence level in (0,1).}

\item{...}{Arguments to pass to \code{fcov}.}
}
\description{
This function computes approximate confidence intervals for the expected response as well as prediction intervals for a nonlinear regression model. It is designed to behave similarly to the \code{predict} function.
}
\details{
Standard errors for confidence intervals are estimated using the delta method. Derivatives are computed numerically rather than analytically to permit a wider range of models. Prediction intervals assume normally-distributed responses with variance inversely proportional to any specified weights, or homoscedastic error if no weights are specified.
}
\examples{
myreg <- nls(rate ~ (t1 + t3 * (state == "treated")) * conc /
 (t2 + t4 * (state == "treated") + conc), data = Puromycin,
 start = c(t1 = 150, t2 = 0, t3 = 0.05, t4 = 0), data = Puromycin) 
nlsint(myreg, interval = "confidence")
}
