% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skylark.R
\docType{data}
\name{skylark}
\alias{skylark}
\title{Impact of pesticides on skylark reproductivity.}
\format{A data frame with 16 observations and four variables:
\describe{
   \item{year}{year on which the observation was made}
   \item{field}{identifier for the field}
   \item{spray}{was the field sprayed that year?}
   \item{count}{number of skylark fledglings}
}}
\source{
Tjur, T. (1998). Nonlinear regression, quasi likelihood, and overdispersion in generalized linear models. \emph{American Statistician}, \emph{52}, 222-227.
}
\usage{
skylark
}
\description{
The data are from a cross-over design in which four fields were either sprayed or not on each of four consecutive years. For each field on each year, the number of skylark fledglings produced was recorded.
}
\keyword{datasets}
